Move length constants to opengever.base.model.
