+++ opengever/activity/model/watcher.py
@@ -1,7 +1,7 @@
 from opengever.base.model import Base
+from opengever.base.model import USER_ID_LENGTH
 from opengever.ogds.base.actor import Actor
 from opengever.ogds.base.actor import ActorLookup
-from opengever.ogds.models import USER_ID_LENGTH
 from sqlalchemy import Column
 from sqlalchemy import Integer
 from sqlalchemy import String
