+++ opengever/base/model/__init__.py
@@ -12,13 +12,19 @@ import sqlalchemy_utils
 DEFAULT_LOCALE = 'de'
 SUPPORTED_LOCALES = ['de', 'fr', 'en']
 
-
 CONTENT_TITLE_LENGTH = 255
-PORTAL_TYPE_LENGTH = 100
-ZIP_CODE_LENGTH = 16
 CSS_CLASS_LENGTH = 100
+EMAIL_LENGTH = 255
+FIRSTNAME_LENGTH = 255
+GROUP_ID_LENGTH = 255
+GROUP_TITLE_LENGTH = 255
+LASTNAME_LENGTH = 255
+PORTAL_TYPE_LENGTH = 100
 UID_LENGTH = 36
-
+UNIT_ID_LENGTH = 30
+UNIT_TITLE_LENGTH = 255
+USER_ID_LENGTH = 255
+ZIP_CODE_LENGTH = 16
 
 tables = [
     'admin_units',
