+++ opengever/contact/contact.py
@@ -1,8 +1,8 @@
 from collective import dexteritytextindexer
+from opengever.base.model import EMAIL_LENGTH
+from opengever.base.model import FIRSTNAME_LENGTH
+from opengever.base.model import LASTNAME_LENGTH
 from opengever.contact import _
-from opengever.ogds.models import EMAIL_LENGTH
-from opengever.ogds.models import FIRSTNAME_LENGTH
-from opengever.ogds.models import LASTNAME_LENGTH
 from plone.dexterity.content import Item
 from plone.indexer import indexer
 from plone.namedfile.field import NamedImage
