+++ opengever/contact/models/archive.py
@@ -1,6 +1,6 @@
 from opengever.base.date_time import utcnow_tz_aware
+from opengever.base.model import USER_ID_LENGTH
 from opengever.base.model import UTCDateTime
-from opengever.ogds.models import USER_ID_LENGTH
 from sqlalchemy import Column
 from sqlalchemy import ForeignKey
 from sqlalchemy import Integer
