+++ opengever/contact/models/archivedmailaddress.py
@@ -1,7 +1,7 @@
 from opengever.base.model import Base
 from opengever.base.model import CONTENT_TITLE_LENGTH
+from opengever.base.model import EMAIL_LENGTH
 from opengever.contact.models.archive import ArchiveMixin
-from opengever.ogds.models import EMAIL_LENGTH
 from sqlalchemy import Column
 from sqlalchemy import Integer
 from sqlalchemy import String
