+++ opengever/contact/models/archivedperson.py
@@ -1,7 +1,7 @@
 from opengever.base.model import CONTENT_TITLE_LENGTH
+from opengever.base.model import FIRSTNAME_LENGTH
+from opengever.base.model import LASTNAME_LENGTH
 from opengever.contact.models.archivedcontact import ArchivedContact
-from opengever.ogds.models import FIRSTNAME_LENGTH
-from opengever.ogds.models import LASTNAME_LENGTH
 from sqlalchemy import Column
 from sqlalchemy import ForeignKey
 from sqlalchemy import Integer
