+++ opengever/contact/models/mailaddress.py
@@ -1,9 +1,9 @@
 from opengever.base.model import Base
 from opengever.base.model import CONTENT_TITLE_LENGTH
 from opengever.base.model import create_session
+from opengever.base.model import EMAIL_LENGTH
 from opengever.base.utils import to_safe_html
 from opengever.contact.docprops import MailAddressDocPropertyProvider
-from opengever.ogds.models import EMAIL_LENGTH
 from sqlalchemy import Column
 from sqlalchemy import ForeignKey
 from sqlalchemy import Integer
