+++ opengever/contact/models/participation.py
@@ -1,14 +1,14 @@
 from opengever.base.model import Base
 from opengever.base.model import create_session
 from opengever.base.model import SQLFormSupport
+from opengever.base.model import UNIT_ID_LENGTH
+from opengever.base.model import USER_ID_LENGTH
 from opengever.base.oguid import Oguid
 from opengever.contact import _
 from opengever.contact.models.org_role import OrgRole
 from opengever.contact.models.participation_role import ParticipationRole
 from opengever.contact.ogdsuser import OgdsUserToContactAdapter
 from opengever.ogds.base.utils import ogds_service
-from opengever.ogds.models import UNIT_ID_LENGTH
-from opengever.ogds.models import USER_ID_LENGTH
 from sqlalchemy import Column
 from sqlalchemy import ForeignKey
 from sqlalchemy import Integer
