+++ opengever/contact/models/person.py
@@ -1,10 +1,10 @@
 from opengever.base.model import CONTENT_TITLE_LENGTH
+from opengever.base.model import FIRSTNAME_LENGTH
+from opengever.base.model import LASTNAME_LENGTH
 from opengever.contact.docprops import PersonDocPropertyProvider
 from opengever.contact.models.contact import Contact
 from opengever.contact.utils import get_contactfolder_url
 from opengever.contact.wrapper import PersonWrapper
-from opengever.ogds.models import FIRSTNAME_LENGTH
-from opengever.ogds.models import LASTNAME_LENGTH
 from sqlalchemy import Column
 from sqlalchemy import ForeignKey
 from sqlalchemy import Integer
