+++ opengever/globalindex/model/task.py
@@ -3,17 +3,17 @@ from DateTime import DateTime as ZopeDateTime
 from opengever.base.model import Base
 from opengever.base.model import is_oracle
 from opengever.base.model import Session
+from opengever.base.model import UNIT_ID_LENGTH
+from opengever.base.model import USER_ID_LENGTH
 from opengever.base.oguid import Oguid
+from opengever.base.query import BaseQuery
+from opengever.base.types import UnicodeCoercingText
 from opengever.base.utils import escape_html
 from opengever.globalindex.model import WORKFLOW_STATE_LENGTH
 from opengever.globalindex.model.reminder_settings import ReminderSetting
 from opengever.ogds.base.actor import Actor
 from opengever.ogds.base.utils import get_current_admin_unit
 from opengever.ogds.base.utils import ogds_service
-from opengever.ogds.models import UNIT_ID_LENGTH
-from opengever.ogds.models import USER_ID_LENGTH
-from opengever.base.query import BaseQuery
-from opengever.base.types import UnicodeCoercingText
 from opengever.task.reminder.reminder import TaskReminder
 from plone import api
 from Products.CMFPlone.utils import safe_unicode
