+++ opengever/locking/model/locks.py
@@ -1,8 +1,8 @@
 from datetime import timedelta
 from opengever.base.date_time import utcnow_tz_aware
 from opengever.base.model import Base
+from opengever.base.model import USER_ID_LENGTH
 from opengever.base.model import UTCDateTime
-from opengever.ogds.models import USER_ID_LENGTH
 from sqlalchemy import Column
 from sqlalchemy import Integer
 from sqlalchemy import String
