+++ opengever/meeting/model/agendaitem.py
@@ -1,7 +1,9 @@
 from AccessControl import getSecurityManager
 from opengever.base.model import Base
 from opengever.base.model import create_session
+from opengever.base.model import UNIT_ID_LENGTH
 from opengever.base.oguid import Oguid
+from opengever.base.types import UnicodeCoercingText
 from opengever.base.utils import to_html_xweb_intelligent
 from opengever.document.interfaces import ICheckinCheckoutManager
 from opengever.globalindex.model import WORKFLOW_STATE_LENGTH
@@ -13,8 +15,6 @@ from opengever.meeting.model.excerpt import Excerpt
 from opengever.meeting.workflow import State
 from opengever.meeting.workflow import Transition
 from opengever.meeting.workflow import Workflow
-from opengever.ogds.models import UNIT_ID_LENGTH
-from opengever.base.types import UnicodeCoercingText
 from opengever.trash.trash import ITrashable
 from opengever.trash.trash import ITrashed
 from plone import api
