+++ opengever/meeting/model/committee.py
@@ -1,6 +1,9 @@
 from datetime import date
 from opengever.base.model import Base
+from opengever.base.model import GROUP_ID_LENGTH
+from opengever.base.model import UNIT_ID_LENGTH
 from opengever.base.oguid import Oguid
+from opengever.base.types import UnicodeCoercingText
 from opengever.base.utils import escape_html
 from opengever.globalindex.model import WORKFLOW_STATE_LENGTH
 from opengever.meeting import _
@@ -12,9 +15,6 @@ from opengever.meeting.workflow import State
 from opengever.meeting.workflow import Transition
 from opengever.meeting.workflow import Workflow
 from opengever.ogds.base.utils import ogds_service
-from opengever.ogds.models import GROUP_ID_LENGTH
-from opengever.ogds.models import UNIT_ID_LENGTH
-from opengever.base.types import UnicodeCoercingText
 from sqlalchemy import and_
 from sqlalchemy import Column
 from sqlalchemy import Integer
