+++ opengever/meeting/model/generateddocument.py
@@ -1,7 +1,7 @@
 from opengever.base.model import Base
+from opengever.base.model import UNIT_ID_LENGTH
 from opengever.base.oguid import Oguid
 from opengever.locking.lock import SYS_LOCK
-from opengever.ogds.models import UNIT_ID_LENGTH
 from plone.locking.interfaces import ILockable
 from sqlalchemy import Column
 from sqlalchemy import Integer
