+++ opengever/meeting/model/proposal.py
@@ -1,5 +1,8 @@
 from opengever.base.model import Base
+from opengever.base.model import UNIT_ID_LENGTH
+from opengever.base.model import USER_ID_LENGTH
 from opengever.base.oguid import Oguid
+from opengever.base.types import UnicodeCoercingText
 from opengever.base.utils import escape_html
 from opengever.document.interfaces import ICheckinCheckoutManager
 from opengever.globalindex.model import WORKFLOW_STATE_LENGTH
@@ -13,9 +16,6 @@ from opengever.meeting.workflow import State
 from opengever.meeting.workflow import Transition
 from opengever.meeting.workflow import Workflow
 from opengever.ogds.base.utils import ogds_service
-from opengever.ogds.models import UNIT_ID_LENGTH
-from opengever.ogds.models import USER_ID_LENGTH
-from opengever.base.types import UnicodeCoercingText
 from plone import api
 from sqlalchemy import Column
 from sqlalchemy import Date
