+++ opengever/ogds/base/browser/team_forms.py
@@ -2,11 +2,11 @@ from ftw.keywordwidget.widget import KeywordFieldWidget
 from ftw.keywordwidget.widget import KeywordWidget
 from opengever.base.browser.modelforms import ModelAddForm
 from opengever.base.browser.modelforms import ModelEditForm
+from opengever.base.model import UNIT_TITLE_LENGTH
 from opengever.contact.utils import get_contactfolder_url
 from opengever.ogds.base import _
-from opengever.ogds.base.sources import CurrentAdminUnitOrgUnitsSourceBinder
 from opengever.ogds.base.sources import AllFilteredGroupsSourceBinder
-from opengever.ogds.models import UNIT_TITLE_LENGTH
+from opengever.ogds.base.sources import CurrentAdminUnitOrgUnitsSourceBinder
 from opengever.ogds.models.team import Team
 from plone.autoform import directives as form
 from plone.autoform.widgets import ParameterizedWidget
