+++ opengever/ogds/base/interfaces.py
@@ -1,4 +1,4 @@
-from opengever.ogds.models import UNIT_ID_LENGTH
+from opengever.base.model import UNIT_ID_LENGTH
 from zope import schema
 from zope.interface import Attribute
 from zope.interface import Interface
