+++ opengever/ogds/base/sync/ogds_updater.py
@@ -1,12 +1,12 @@
 from logging.handlers import TimedRotatingFileHandler
 from opengever.base.model import create_session
+from opengever.base.model import GROUP_ID_LENGTH
+from opengever.base.model import USER_ID_LENGTH
 from opengever.base.pathfinder import PathFinder
 from opengever.ogds.base.interfaces import ILDAPSearch
 from opengever.ogds.base.interfaces import IOGDSSyncConfiguration
 from opengever.ogds.base.interfaces import IOGDSUpdater
 from opengever.ogds.base.sync.import_stamp import set_remote_import_stamp
-from opengever.ogds.models import GROUP_ID_LENGTH
-from opengever.ogds.models import USER_ID_LENGTH
 from opengever.ogds.models.group import Group
 from opengever.ogds.models.user import User
 from plone import api
