--- opengever/ogds/models/__init__.py
@@ -1,8 +1,0 @@
-UNIT_ID_LENGTH = 30
-UNIT_TITLE_LENGTH = 255
-GROUP_ID_LENGTH = 255
-USER_ID_LENGTH = 255
-FIRSTNAME_LENGTH = 255
-LASTNAME_LENGTH = 255
-EMAIL_LENGTH = 255
-GROUP_TITLE_LENGTH = 255
