+++ opengever/ogds/models/admin_unit.py
@@ -1,6 +1,6 @@
 from opengever.base.model import Base
-from opengever.ogds.models import UNIT_ID_LENGTH
-from opengever.ogds.models import UNIT_TITLE_LENGTH
+from opengever.base.model import UNIT_ID_LENGTH
+from opengever.base.model import UNIT_TITLE_LENGTH
 from opengever.ogds.models.group import groups_users
 from opengever.ogds.models.org_unit import OrgUnit
 from opengever.ogds.models.user import User
