+++ opengever/ogds/models/org_unit.py
@@ -1,10 +1,10 @@
 from opengever.base.model import Base
-from opengever.ogds.models import GROUP_ID_LENGTH
-from opengever.ogds.models import UNIT_ID_LENGTH
-from opengever.ogds.models import UNIT_TITLE_LENGTH
+from opengever.base.model import GROUP_ID_LENGTH
+from opengever.base.model import UNIT_ID_LENGTH
+from opengever.base.model import UNIT_TITLE_LENGTH
+from opengever.base.query import BaseQuery
 from opengever.ogds.models.group import Group
 from opengever.ogds.models.inbox import Inbox
-from opengever.base.query import BaseQuery
 from sqlalchemy import Boolean
 from sqlalchemy import Column
 from sqlalchemy import ForeignKey
