+++ opengever/ogds/models/team.py
@@ -1,8 +1,8 @@
-from opengever.base.model import SQLFormSupport
 from opengever.base.model import Base
-from opengever.ogds.models import GROUP_ID_LENGTH
-from opengever.ogds.models import UNIT_ID_LENGTH
-from opengever.ogds.models import UNIT_TITLE_LENGTH
+from opengever.base.model import GROUP_ID_LENGTH
+from opengever.base.model import SQLFormSupport
+from opengever.base.model import UNIT_ID_LENGTH
+from opengever.base.model import UNIT_TITLE_LENGTH
 from opengever.base.query import BaseQuery
 from sqlalchemy import Boolean
 from sqlalchemy import Column
