+++ opengever/ogds/models/user.py
@@ -1,8 +1,8 @@
 from opengever.base.model import Base
-from opengever.ogds.models import EMAIL_LENGTH
-from opengever.ogds.models import FIRSTNAME_LENGTH
-from opengever.ogds.models import LASTNAME_LENGTH
-from opengever.ogds.models import USER_ID_LENGTH
+from opengever.base.model import EMAIL_LENGTH
+from opengever.base.model import FIRSTNAME_LENGTH
+from opengever.base.model import LASTNAME_LENGTH
+from opengever.base.model import USER_ID_LENGTH
 from opengever.base.query import BaseQuery
 from opengever.base.types import UnicodeCoercingText
 from sqlalchemy import Boolean
